backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.28
psycopg[binary]==3.1.18

alembic==1.13.1

pyjwt==2.8.0
cryptography>=44.0.0

httpx==0.27.0

# Google API clients
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
google-cloud-firestore>=2.14.0
google-cloud-datastore>=2.18.0
google-cloud-storage>=2.13.0

python-dotenv==1.0.1
pydantic>=2.12.2

email-validator==2.1.0.post1
black==24.3.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-env==1.1.3
slowapi==0.1.9
